Add extraction of revisions
With this patch, the list of revisions as stored in the database is
extracted, too. The list is stored in the file
"{project_dir}/revisions.list".

If the list of revisions as given with the project configuration file
does not match the list inside the DB, the extraction process is
stopped.
If the list of revisions in the configuration file is empty, the list
from the DB is used instead.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>