Change behaviour of issue analysis
Instead of starting a java programm to do the issue processing
in the issue part of codeface-extraction, the processing is now
performed in python and uses a json file containing the issues. Also
this adds reference target entry to the output list when a user is
mentioned and subscribed in a comment.

Signed-Off-By: Raphael Nömmer <R.Noemmer@gmx.de>