Extend mbox parsing to search for files
Add the '--file' CLI flag to search for file names occurring in e-mails of
the mbox.

Notice: Using '--file' just searches for base names of files, *not* for
the whole file path. To search for the whole file path of a file (which is
its unique identifier), you have to use both the '--file' flag and the
'--filepath' flag.

This fixes #9.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>