Adaptively load JIRA file with persons information
As Codeface changed the output file containing JIRA person information
in the conway analysis [1], we need to adaptively load the correct file.
Therefore, we try to load the following file names in the given order
and only throw an error when none exists on disk:
- 'jira-comment-authors-with-email.csv', and
- 'jira_issue_comments.csv'.

Some obsolete comments are removed.

[1] https://github.com/siemens/codeface/commit/5ba4c1bc8b60fe (line 136)