apply review on previous commits
The review from clhunsen and bockthom on the previous two commits gets
applied. Especially following things change: The 'with_api' flag is
renamed to 'skip_history' and is now part of the  command-line API. The
issue-url is now loaded from the configuration file. The information
order in the output-file changed. Now there are fields for the global
state and resolution and fields for the state and resolution on
creation. Issue state and resolution on creation now has default values.
The API informations get adjust to the XML-informations. The copyrigth
header gets updated.

Signed-off-by: Anselm Fehnker <fehnker@fim.uni-passau.de>