Refactor function "remove_problematic_characters" to deal better with encoding
Fixed function in order to correctly handle the encoding applied by
codeface prior to writing the contents to the data base. The fix is
specifically for content that is originally encoded in utf-8 and is
tailored to the unicode conversion in codeface (function
"encode_as_utf8").

NOTE: The fix might work as expected for content that stems from other
than utf-8 encoded source files. Known errors with non-utf-8 encoded
source files: Any umlauts will be removed.

Signed-off-by: Tina Schuh <schuht@fim.uni-passau.de>