Fix log output when not performing range-level extractions
When range-level extractions are *not* enabled (the default), the
statement "Extracting data for version '{hash}'" still appeared in the
log output. As no range data is extracted, also the log statement should
not be printed. This commit fixes that and also the list of revisions is
only checked when the range-level extractions are enabled.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>