Adjust e-mail extraction to handle multiple mailing lists
For the case that there are several projects configured in a Codeface
configuration, the mailing lists are stored with a unique ID but the
thread IDs of the individual e-mails are only unique to their respective
mailing list.

To circumvent the problem, the e-mail extraction ('EmailExtraction' and
'EmailRangeExtraction' classes) now construct a unique thread ID of the
following format: "{mailing-list ID}#{old thread ID}".

Note: This does not break anything with the network library as the
thread ID (old or new, does not matter) is handled as a String constant
– what still works although the individual Strings may change for all
future extractions.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>