Change argparse names to be more consistent
The parameterst 'prog' for the 'ArgumentParser' instances are all set to
'codeface' – which is obviously wrong as this is the codeface-extraction
project. Change the given arguments to be more consistent with the
scripts in which they are constructed.

This is no functional change or even fix, this just helps to prevent
misunderstanding the log output in the case of errors due to missing
arguments, for example.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>