Create one index per mailing list in mbox parsing
Curently, the indexing during the mbox parsing does not work for
multiple mailing lists. In detail, the first run is performed correctly
(or when setting '--reindex'), but for the next run (without
'--reindex') only the index for the last mbox is present. This results
in incomplete results.

The problem is fixed by creating an index for each individual mailing
list in a project. Inside the project's result folder, all indices are
kept in folders 'mbox-index/{ml_name}/'. For this to work, the mbox name
is passed to the function '__get_index'.

For debugging purposes, the removal of an index is logged appropriately.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>