Adjust range directories' names
To conform with the names of range directories created by codeface, the
range directories generated throughout the extraction are adjusted to
the names used by codeface. That is, the commit hashes forming the range
are shortend to a length of six characters and prepended by a
consecutive range number.

See siemens/codeface@cd7b68c65ff7ae113e6c75275ce3798004ce7b09.

This fixes se-passau/codeface-extraction#21.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>