Fix name encoding when extracting author, commit, and e-mail data
In the database, there are sometimes ill-encoded names originating from
e-mail data using specific header encodings. When extracting author,
commit, and e-mail data from the database, we still extract those
ill-encoded names.

In this commit, we post-process names after extracting them from the
database and fix the encoding if possible, to get well-readable names in
the resulting .list files.

Props to @nischikata for making us aware of how this can be fixed and
for providing a suitable draft of the eventual solution.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>