Check content type in mbox-parsing only for lower-case characters
In the mbox-parsing, when checking whether the content-type of a message
contains "text", the content type needs to be set to lower case
characters before performing the check for "text" being part. Reason:
Usually the content type is, e.g., "text/plain", but there are lots of
e-mails whose content type is "TEXT/PLAIN".

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>