Unify GitHub users
When reformating the events of GitHub issues, update and unify users:
For instance, we can get actual names and e-mail addresses from
referenced commits for users for which we only have a username. To add
this information also for other events where those data is missing, a
dictionary of users is built and if some data is missing, the data from
the dictionary is used.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>