Fix mention and subscribe events
Currently, we assume that mention and subscribe events take place at
exactly the same time as the comment in which they get mentioned.
However, there are examples (e.g., issue number 8834 of OpenSSL) in
which the mention and subscribe events take place one seconds later than
the actual comment they get mentioned in. Hence, if we do not find a
comment at exactly the same time, we have to look for a comment one
second earlier to be able to correctly fix the mentioned and subscribe
events' authors.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>