Create dictionary of users from all issues
It is not enough to create the users dictionary just from the events of
one issue, as a user can only be active in another issue and will not
have it in the dictionary of an issue where the user is referenced.
Hence, the dictionary is created upon all issues.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>