Remove duplicated events
When someone merges a pull request, the merge commit originating from
the merge of the pull request causes two "add_link" commits, resulting
in two "add_link" events of the same user at exactly the same time
referencing the exactly the same commit in exactly the same pull
request. As there are two events for referencing the same commit, we
need to remove duplicates (but preserve the order of the events).

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>