Fix review information of GitHub pull requests
When a review on GitHub gets dismissed in a later event, the original
state of the review gets replaced by "dismissed". Hence, the state of
the review is wrong as it was not "dismissed" at the moment the review
was created. However, the "review_dismissed" event stores the original
state of the dismissed review. As a consequence, we have to replace the
state of the review by its original state. (To not omit the information
that the review was dismissed later on, the "event_info_2" field is set
to "later_dismissed".)

In addition, in the case that a review is manually dismissed, the person
who dismisses the review can provide a short dismissal message. If such
a message is available, we add an additional "commented" event for that
to the list of comments.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>