Use user information from authors of commits on GitHub
As the json file resulting from the GitHubWrapper does contain author
and committer information for all related commits, we can use this
information also for building our user dictionary, which is necessary to
merge GitHub usernames with real names and e-mail addresses resulting
from the git repository. As the actor of a related-commit event is the
committer in general, we only need to additionally add the author data.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>