Add additional "commented" event for the creation event of a GitHub issue
Adding an additional "commented" event for the creation of an issue is
necessary as a the creation comment can already mention other users. As
only "commented" events are used to fix authors and targets of
"mentioned" and "subscribe" events, we need to add the creation comment
also to the list of comments and, therefore, we add an additional
"commented" event.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>