Add "referenced_by" event to each issue which gets referenced by another issue
The list of related issues contains events in which a user has
referenced another issue within the current issue. In this commit, the
reversed event is also added: The referenced issue gets an event which
states that is was referenced in another issue.

As we cannot modify another issue while iterating over the list of
issues, the reversed events need to be temporarily stored in a temporary
list and added to the actual issue afterwards.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>