Distinguish whether a commit references an issue or the other way round
The event "add_link" is used when someone references another issue or
another commit in the current issue (e.g., in a comment).

The event "referenced_by" is used when someone references the current
issue within a commit message or in another issue.

For related issues, this was already fixed previously. Now, this is also
fixed for related commits, to stay consistent.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>