Fix removing types and resolutions of GitHub issues
In some strange cases, in a GitHub issue, somebody has removed a label
that was never added before. If the removed label also is a "type" or an
"resolution", removing this type or resolution from the corresponding
list does not work as it is not in the list. So, only remove it if it is
in the list.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>