Fix undefined variable in 'CommitRangeExtraction'
The variable 'committerDate' is not defined in the context of
'CommitRangeExtraction::_reduce_result'. When enabling '--range' in the
CLI, an error is produced and the extraction crashes. This is a
regression introduced by commit d95967f7e18b367359d417fc3cdb51f08a04a85b
and which has been missed in the review for PR #23.

The mistake is fixed by accessing the wanted variable 'commitDate'.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>