Retrieve each commit only once in the commit extraction
As Codeface analyses the data in three-month ranges, commits which are
on the border of those ranges are included twice as they belong to two
different ranges and, therefore, end up with two different commit ids in
the database though having the same commit hash.

(It is not enough to only filter those commits based on the commit
hashes describing the start or end of a Codeface range. The reason for
that is that Codeface assigns commits to a range not based on the commit
hash describing the start/end of a range, but on the commit date (not
the author date) of the belonging to the commit hash which desribes the
start/end. Hence, all those commits which have the same commit date or
also assigned to both ranges!)

To mitigate this, only extract such commits once, that is, just take the
entries having the lower commit id of those commits having the same
hashes.

(It is not necessary to adjust the range extraction as within each
range, each commit appears already only once. However, keep in mind that
when combining all the range extractions, the commits at the border will
still be duplicated.)

This fixes #28.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>