Remove superfluous join from range level extractions
As we do not need the release-timeline data of the previous range in
range-level extractions on commits, we can remove the corresponding JOIN
in the SQL statements.

Thanks to @clhunsen for pointing this out.

In addition, add a short comment on range bounds: Ranges include their
ends as Codeface uses the committer date of a commit to split the ranges.
Commits at the end may, therefore, have an earlier author date, which is
the reason that such commits better suit into the previous range than the
subsequent range.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>