Fix problems with user "null" in GitHub issue data
In very rare cases of the reviews API of GitHub the user is "null" where
it should be the ghost user (which replaces deleted users). When parsing
the resulting json file, the "null" object does not have a name or
e-mail address, resulting in an exception when name or e-mail address
are accessed.

To fix this, the user "null" gets replaced by the ghost user (deleted
user) as soon as its name or e-mail addressed are accessed.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>