JIRA issue processing: Don't run gephi export and archaical issues export
As the gephi export is not used and also not available for other extractions
or issue processing steps, don't run the gephi export anymore.
Same holds for the archaic issues export: Don't run it anymore as it is not
used and not compatible to other tools like coronet.

As these two exports are just commented out, still create empty files for them
at the beginning to avoid getting wrong results when someone wants to run these
commented exports but omitting clearing the corresponding output files.

Props to @ecklbarb for discussing whether the above mentioned exports
are still needed.

This is related to issue #33.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>
Signed-off-by: Barbara Eckl-Ganser <ecklbarb@fim.uni-passau.de>