Add backup option for author postprocessing
Add an additional CLI parameter to enable data backups before performing
the postprocessing. If data backups are enabled, all the  files of the
results directory are copied to another results directory having
suffix '_bak'. If this results directory does already exists, no backup
is performed even if backups are enabled.

Suggested-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>