Fix problems with user buffer in issue processing
The user buffer used in the issue processing contained outdated
information as user updates in the DB have not been considered any more.
To fix this, we need two iterations over all the issue events:

(1) In the first iteration, we just update the user information in the
    database after each event and just add the person's id from the
    database into the buffer (instead of a concrete name-email pair).
    In the issue data, we replace the user by its id from the database.

(2) In the second iteration, for each issue event, we get the user's
    person id from the buffer and extract the current name-email pair
    from the database (which is correct now as any possible update
    has already been performed in the first iteration) and add this
    one to a new buffer (whose key is the person id). Then we can
    lookup the name-email pair based on the person id in the buffer.

Afterwards, all the buffers are up to date and we don't use outdated
information any more.

This fixes #34.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>