Fix bug in author postprocessing of 'event.info.2'
The column 'event.info.2' (column 13) does not contain a single e-mail
address but was generated during issue processing by calling
`json.dump()` on a list containing just one element.

Hence, if 'event.info.2' contains an e-mail address, the e-mail address
is surrounded by quotation marks within the string. As a consequence,
we need to add those quotation marks to the e-mail address when checking
for equality with column 13 and also we need to add those quotation marks
when re-assigning elements of column 13.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>