Deal with missing author name in extraction
When we don't have an author name for an author entry in the person table of the
codface database, the name actually is 'None'. However, 'None' cannot be encoded
into 'utf-8', resulting in an error. Therefore, if the name is 'None', just
return the name as is, without trying to encode the name.

Note: This results in an empty string in the "author name" column of the
corresponding '*.list' files. As an author without name might be a dummy user,
take care of the author whose name is "" when processing and using the resulting
'*.list' files!

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>