Improve merging of user data and JIRA issue data
Up to now, we merged usernames in the JIRA issue data with person data
from JIRA by username. However, if we don't have a username, no merging
was performed before. But, in this case, we can try to merge by name
instead. Therefore, add a second dictionary which stores the users by
name, so that we can merge either by name or by username, depending on
which information is given.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>