Get persons' names from JIRA history
When extracting events from JIRA history, we only extracted usernames,
but we can also get names (display names). In the case that the
username is not in our person data, but the name is, we can make use of
the name. Therefore, also collect the names when extracting users from
JIRA history.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>