Fix user updates to database in JIRA issue processing
When updating the users with the Codeface database, we only considered
issue authors and issue-comment authors. However, when processing JIRA
data, we also need to process the history data, that is, authors of
events and assigned users. In this commit, we fix this behavior and
update also the users occurring in the history data with the Codeface
database.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>