Handle JIRA Error during history extraction
When extracting changelog information for a JIRA issue via the
corresponding API, it can happen that the changelog information for this
issue is not available (for whatever reason) und cannot be extracted. In
such a case, the API call ends up in a JIRA Error. If this happens, just
catch this error, print a warning, and create an empty history, so that
no further events are processed in this case.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>