[bot processing] Enable adding bots from project-independent list of known bots
We now can have a global, project-independent list of known bots. This file has
the name "known_github_bots.list" and is located in the "configurations"
directory. The file just has to contain **GitHub usernames**, one username per
row. All usernames in this file are treaded as bots then.

When processing the bot data we have extracted from a project, we check whether
also any of these known bots does occur in the GitHub issue data of the project.
If so, we add the known bot to the list of bots. If the username was already
present in the list of bots before, the prediction value is updated accordingly.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>