Properly handle author 'GitHub <noreply@github.com>'
The author 'GitHub <noreply@github.com>' can appear in the data, e.g., when the
Web frontend of GitHub is used for editing files. However, this distorts the
data as the actual user is obfuscated. To avoid problems originating from this
"author", we need to get rid of this author. We do this in the following way:

(1) Commit data:
Commits that originate from editing files in the Web frontend of GitHub have
their actual author as author and 'GitHub <noreply@github.com>' as committer. As
the author who edits the file should also be the committer in this situation, we
replace the committer 'GitHub <noreply@github.com>' by the author of the commit.

(2) Issue data:
The author of "commit_added" events usually is the committer of the added
commit. As we replace the committer 'GitHub <noreply@github.com>' in the commit
data by the author of the commit, we also perform this analogously for the
author of the "commit_added" events in the issue data, that is, we extract the
author of the commit from the commit data.

If the added commit (added by 'GitHub <noreply@github.com>') does not occur in
the commit data, we just remove the corresponding "commit_added" event from the
issue data. This mostly (but not only) happens when a pull request contains a
merge commit, as merge commits are not tracked by Codface and, therefore, do not
occur in the commit data.

(3) Author data:
As we remove the author 'GitHub <noreply@github.com>' from the issue data and
from the commit data, we also need to remove this user from the author data, to
keep the data consistent.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>