Properly handle author 'GitHub <noreply@github.com>' for mail data
As a follow up for commit 8064b0d1083523bcec6ba442cfeeb26cc6bc95d2, remove the
author 'GitHub <noreply@github.com>' also from the e-mail data, that is, remove
the e-mails that are sent by this author.

As Codeface sometimes adjusts the e-mail address of this user, handling this
author is performed analogously on 'GitHub <GitHub.noreply@github.com>'.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>