Properly handle author 'GitHub <noreply@github.com>' for GitHub issue data
In the issue data, up to now, we assumed that the user 'GitHub
<noreply@github.com>' only occurs as author of "commit_added" events. However,
it turned out that this user also can occur in "mentioned" or "subscribed"
events, for instance. As we completely ignore this user, we have to treat this
user in the following way when postprocessing the GitHub issue data:

- For "commit_added" events, replace the user 'GitHub <noreply@github.com>' by
the author of the commit. If this is not possible, just remove the event. (This
was already previously implementend.)
- For all other events, if the author of the event is 'GitHub
<noreply@github.com>', just remove the event.
- For all other "mentioned" or "subscribed" events, if the referenced user is
'GitHub <noreply@github.com>', just remove the event.

This way, the user 'GitHub <noreply@github.com>' should not appear in the GitHub
issue data any more.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>