Fix overwriting usernames.list by changing the directory for this file
When performing the GitHub issue processing, we write a file 'usernames.list'
which contains a mapping from usernames to name/e-mail pairs, needed for bot
processing. However, in the previous implementation, we stored this file in
the '_issues' directory, independent of the tagging ('proximity' or 'feature'),
ending up in feature and proximity analysis overwriting the file of each other.

As feature and proximity analsis can lead to different e-mail addresses
(especially in randomly generated e-mail addresses if no e-mail address is
available), we need to store them separatly for feature and proximity analysis,
otherwise bot processing will fail.

With this commit, we change the directory where the 'usernames.list' list file is
stored from '_issues' to the '_proximity/proximity' and '_feature/feature'
directory respectively.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>