Remove duplicates from bot data
It can happen that authors are joined multiple times to the "usernames.list".
Therefore, when processing the bot data, we need to remove duplicates.

Again, this can also happen when post-processing the bot data: When replace the
(author.name, e-mail) pair in the bot data according to the disambigation list,
a (author.name, e-mail) pair might be introduced that has already been
introduced before. So, here, we also need to remove duplicates (and make sure
that we keep the correct entry (e.g., where the last column is "Bot" and not
"unknown", if we have multiple entries with different classifications).
We keep "Bot" if, at least, one of the classifications is "Bot". Otherwise, we
keep "Human" if, at least, one of the classifications is "Human".

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>