Do not try to fix encoding of non-existent text
When the commit message of a commit is missing, then `fix_characters_in_string`
is called with parameter `None`. However, `None` is not allowed in the function
`fix_encoding`. Therefore, we need to perform a `None` check before
`fix_encoding` is called. If the text input is `None`, we now just return `None`
without any other fixes.

This change is caused by the new handling of missing commit messages in
codeface, as now there can be commits without commit message in the codeface
database. For more information, see also:
https://github.com/se-sic/codeface/commit/51073b2d86448b504689f0cb43d99b2007911d03

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>